use CreateEventPage instead of CopyEventPage (test does not work yet)
