+++ src/domain/event/CreateEventPage.tsx
@@ -1,53 +1,132 @@
+import { useApolloClient } from '@apollo/react-hooks';
 import { Notification } from 'hds-react';
-import { compact } from 'lodash';
-import * as React from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useHistory } from 'react-router';
+import { useSelector } from 'react-redux';
+import { useHistory, useLocation, useParams } from 'react-router';
 import { toast } from 'react-toastify';
 
 import LoadingSpinner from '../../common/components/loadingSpinner/LoadingSpinner';
 import {
-  OrganisationNodeFieldsFragment,
+  EventDocument,
+  EventQuery,
+  PersonFieldsFragment,
   useCreateEventMutation,
+  useMyProfileQuery,
+  useUpdateSingleImageMutation,
 } from '../../generated/graphql';
 import useLocale from '../../hooks/useLocale';
-import { Language } from '../../types';
 import { isTestEnv } from '../../utils/envUtils';
 import { clearApolloCache } from '../app/apollo/utils';
 import Container from '../app/layout/Container';
 import PageWrapper from '../app/layout/PageWrapper';
 import { ROUTES } from '../app/routes/constants';
+import { getImageName } from '../image/utils';
+import { getSelectedOrganisation } from '../myProfile/utils';
 import ActiveOrganisationInfo from '../organisation/activeOrganisationInfo/ActiveOrganisationInfo';
-import EventForm, { createEventInitialValues } from './eventForm/EventForm';
-import {
-  useCreateOrUpdateVenueRequest,
-  useUpdateImageRequest,
-} from './eventForm/useEventFormSubmitRequests';
-import useOrganisationPersons from './eventForm/useOrganisationPersons';
+import { activeOrganisationSelector } from '../organisation/selector';
+import { createOrUpdateVenue } from '../venue/utils';
+import EventForm, {
+  createEventAlwaysEmptyInitialValues,
+  createEventInitialValues,
+} from './eventForm/EventForm';
 import styles from './eventPage.module.scss';
 import { CreateEventFormFields } from './types';
 import {
   firstOccurrencePrefilledValuesToQuery,
+  getEventFormValues,
+  getEventLanguageFromUrl,
   getEventPayload,
 } from './utils';
 
-const useEventFormCreateSubmit = (
-  selectedLanguage: Language,
-  selectedOrganisation: OrganisationNodeFieldsFragment | null | undefined
-) => {
+const CreateEventPage: React.FC = () => {
+  const apolloClient = useApolloClient();
+  const { id: eventIdToCopy } = useParams<{ id: string }>();
   const { t } = useTranslation();
-  const locale = useLocale();
   const history = useHistory();
-  const [createEvent] = useCreateEventMutation();
-  const createOrUpdateVenueRequestHandler = useCreateOrUpdateVenueRequest(
-    selectedLanguage
+  const location = useLocation();
+  const locale = useLocale();
+  const language = getEventLanguageFromUrl(location.search);
+  const [selectedLanguage, setSelectedLanguage] = useState(language || locale);
+
+  const [initialValues, setInitialValues] = useState<CreateEventFormFields>(
+    createEventInitialValues
   );
-  const updateImageRequestHandler = useUpdateImageRequest();
+  const [loading, setLoading] = useState<boolean>(true);
+
+  const [createEvent] = useCreateEventMutation();
+  const [updateImage] = useUpdateSingleImageMutation();
+  const { data: myProfileData } = useMyProfileQuery();
+  const handleError = useCallback(
+    (err: Error) => {
+      // TODO: Improve error handling when API returns more informative errors
+      if (isTestEnv()) {
+        // eslint-disable-next-line no-console
+        console.log(err);
+      }
+      toast(t('createEvent.error'), {
+        type: toast.TYPE.ERROR,
+      });
+    },
+    [t]
+  );
+
+  useEffect(() => {
+    const getInitialValues = async () => {
+      try {
+        if (eventIdToCopy) {
+          const { data } = await apolloClient.query<EventQuery>({
+            query: EventDocument,
+            fetchPolicy: 'network-only',
+            variables: {
+              id: eventIdToCopy,
+              include: ['audience', 'in_language', 'keywords', 'location'],
+            },
+          });
+          setInitialValues({
+            ...getEventFormValues(data, selectedLanguage),
+            ...createEventAlwaysEmptyInitialValues,
+          });
+        } else {
+          setInitialValues(createEventInitialValues);
+        }
+      } catch (err) {
+        setInitialValues(createEventInitialValues);
+        handleError(err);
+      } finally {
+        setLoading(false);
+      }
+    };
+    getInitialValues();
+  }, [
+    eventIdToCopy,
+    apolloClient,
+    handleError,
+    selectedLanguage,
+    setInitialValues,
+    setLoading,
+  ]);
+
+  const activeOrganisation = useSelector(activeOrganisationSelector);
+  const selectedOrganisation =
+    myProfileData?.myProfile &&
+    getSelectedOrganisation(myProfileData.myProfile, activeOrganisation);
+
+  const persons =
+    selectedOrganisation?.persons.edges.map(
+      (edge) => edge?.node as PersonFieldsFragment
+    ) || [];
+  const goToEventList = () => {
+    history.push(ROUTES.HOME);
+  };
 
   const handleSubmit = async (values: CreateEventFormFields) => {
     try {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const requests: Promise<any>[] = compact([
+      const requests: Promise<any>[] = [];
+
+      // Request to create new event
+      requests.push(
         createEvent({
           variables: {
             event: {
@@ -60,10 +139,38 @@ const useEventFormCreateSubmit = (
               draft: true,
             },
           },
-        }),
-        createOrUpdateVenueRequestHandler(values),
-        updateImageRequestHandler(values),
-      ]);
+        })
+      );
+
+      const createOrUpdateVenueRequest = createOrUpdateVenue({
+        venueFormData: values,
+        language: selectedLanguage,
+        locationId: values.location,
+      });
+
+      if (createOrUpdateVenueRequest) {
+        requests.push(createOrUpdateVenueRequest);
+      }
+
+      const imageId = values.image;
+      if (imageId) {
+        const imageName = getImageName(imageId);
+        if (imageName) {
+          // Request to update image data
+          requests.push(
+            updateImage({
+              variables: {
+                image: {
+                  altText: values.imageAltText,
+                  id: values.image,
+                  name: imageName,
+                  photographerName: values.imagePhotographerName,
+                },
+              },
+            })
+          );
+        }
+      }
 
       // Run all requests parallel
       const responses = await Promise.all(requests);
@@ -87,48 +194,9 @@ const useEventFormCreateSubmit = (
         search: firstOccurrencePrefilledValuesToQuery(values),
       });
     } catch (e) {
-      // TODO: Improve error handling when API returns more informative errors
-      if (isTestEnv()) {
-        // eslint-disable-next-line no-console
-        console.log(e);
-      }
-      toast(t('createEvent.error'), {
-        type: toast.TYPE.ERROR,
-      });
+      handleError(e);
     }
   };
-
-  return handleSubmit;
-};
-
-const CreateEventPage: React.FC<{
-  initialValues: CreateEventFormFields;
-  loading: boolean;
-  selectedLanguage: Language | undefined;
-  setSelectedLanguage: (selectedLanguage: Language) => void | undefined;
-}> = ({
-  initialValues = createEventInitialValues,
-  loading = false,
-  selectedLanguage,
-  setSelectedLanguage,
-}) => {
-  const { t } = useTranslation();
-  const locale = useLocale();
-  const history = useHistory();
-  const [defaultSelectedLanguage, defaultSetSelectedLanguage] = React.useState(
-    locale
-  );
-  const { organisation, persons } = useOrganisationPersons();
-
-  const onSubmit = useEventFormCreateSubmit(
-    selectedLanguage ?? defaultSelectedLanguage,
-    organisation
-  );
-
-  const goToEventList = () => {
-    history.push(ROUTES.HOME);
-  };
-
   return (
     <PageWrapper title="createEvent.pageTitle">
       <LoadingSpinner isLoading={loading}>
@@ -140,13 +208,11 @@ const CreateEventPage: React.FC<{
             <ActiveOrganisationInfo />
             <EventForm
               onCancel={goToEventList}
-              onSubmit={onSubmit}
+              onSubmit={handleSubmit}
               persons={persons}
               initialValues={initialValues}
-              selectedLanguage={selectedLanguage ?? defaultSelectedLanguage}
-              setSelectedLanguage={
-                setSelectedLanguage ?? defaultSetSelectedLanguage
-              }
+              selectedLanguage={selectedLanguage}
+              setSelectedLanguage={setSelectedLanguage}
               title={t('createEvent.title')}
             />
           </div>
